+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 # [8.16.2] - 2025-02-05
+- Added `disconnected_by` enum to `com.vonage.client.voice.EventWebhook`
 - Changed `AnswerWebhook#getUuid()` return type to String
 - Updated Voice `Call` documentation